deltag: check to see if the user is the user who created the tag OR has a higher permlevel before deleting HAHA YES FINALLY FUCKING FINALLY I DID IT THANK YOU JACK FROM HELPING ME FROM BEYOND THE GRAVE

+++ modules/cmd/deltag.js
@@ -14,7 +14,7 @@ const metadata = {
   long_desc: `Deletes a tag from the tag database. This action **CANNOT BE UNDONE**`,
   args: '<tag name>',
   image: 'IMG_args',
-  authlvl: 3,
+  authlvl: 1,
   flags: ['DM_OPTIONAL', 'LITE'],
   run: null
 };
@@ -26,10 +26,17 @@ metadata.run = (m, args, data) => {
 
   if (!tagName) return OBUtil.err('Missing tag name', { m });
 
-  tagsDB.remove({ name: tagName }, {}, (err, numRemoved) => {
+  tagsDB.find({ name: tagName }, (err, doc) => {
     if (err) return OBUtil.err(err, { m });
 
-    m.channel.send(`✅ \`|\` :pencil: **Tag \`${tagName}\` deleted.**`).then(bm => OBUtil.afterSend(bm, m.author.id));
+    if (m.author.id !== doc[0].userID || OBUtil.getAuthlvl(bot.mainGuild.members.cache.get(doc[0].userID)) >= data.authlvl ) 
+      return OBUtil.err('You cannot delete this user\'s tag!', { m });
+
+    tagsDB.remove({ name: tagName }, {}, (err, numRemoved) => {
+      if (err) return OBUtil.err(err, { m });
+  
+      m.channel.send(`✅ \`|\` :pencil: **Tag \`${tagName}\` deleted.**`).then(bm => OBUtil.afterSend(bm, m.author.id));
+    });
   });
 };
